requirements.txt
gcsa>=2.1.0
requests>=2.28.1
pyyaml>=6.0.1
pytest>=7.2.1
python-dateutil>=2.8.2
functions-framework==3.*
